Added localization for enumerated elements, set test to use English
