+++ Document/Sources/Document/Assets/ru.lproj/Localizable.strings
@@ -24,6 +24,9 @@
 /* Description for the 'adjustable' accessibility trait */
 "trait.adjustable.description" = "Элемент регулировки.";
 
+/* Format for the description of enumerated element of adjustable option */
+"trait.adjustable.enumerated" = "%1$d из %2$d";
+
 /* Hint describing how to use elements with the 'adjustable' accessibility trait */
 "trait.adjustable.hint" = "Смахните вверх или вниз одним пальцем, чтобы настроить значение.";
